Move PoolCandidates table to poolCandidates folder
