+++ admin/resources/js/components/poolCandidate/PoolCandidatesTable.tsx
@@ -1,8 +1,8 @@
 import React, { useMemo } from "react";
 import { Column } from "react-table";
-import { GetPoolCandidatesQuery } from "../api/generated";
-import { notEmpty } from "../helpers/util";
-import Table from "./Table";
+import { GetPoolCandidatesQuery } from "../../api/generated";
+import { notEmpty } from "../../helpers/util";
+import Table from "../Table";
 
 const ClassificationTable: React.FC<GetPoolCandidatesQuery> = ({
   poolCandidates,
